Name eight Common sources after the rows they already hold
Filename-only work, no identity claim: each file holds one function whose
ledger row already spells a real name, and the filename was the mangling of a
promoted body. The six OptionPreferences getters are the ones a player-facing
setting (FPS limit, send delay, replay camera, building occlusion, firewall
refresh) is read through, so they are worth finding by name.

  promoted__getFPSLimitEnabled_*          -> OptionPreferences_getFPSLimitEnabled
  promoted__getSendDelay_*                -> OptionPreferences_getSendDelay
  promoted__saveCameraInReplays_*         -> OptionPreferences_saveCameraInReplays
  promoted__useCameraInReplays_*          -> OptionPreferences_useCameraInReplays
  promoted__getBuildingOcclusionEnabled_* -> OptionPreferences_getBuildingOcclusionEnabled
  promoted__getFirewallNeedToRefresh_*    -> OptionPreferences_getFirewallNeedToRefresh
  promoted__init_CampaignManager_*        -> CampaignManager_init
  promoted__isGroupAiDead_AIGroup_*       -> AIGroup_isGroupAiDead
