+++ Code/GameEngine/Source/Common/OptionPreferences_useCameraInReplays.cpp
@@ -1,0 +1,101 @@
+// cl: /DNDEBUG /DWIN32 /D_WINDOWS /MD /EHsc
+// readable body of ?useCameraInReplays@OptionPreferences@@QAE_NXZ: Code/GameEngine/Source/GameClient/GUI/GUICallbacks/Menus/OptionsMenu.cpp
+// Grok promote from masm_dumps — retail 0x00090900 size 87
+// was: Code/masm_dumps/_sa14__useCameraInReplays_OptionPreferences__QAE_NXZ_90900.asm
+
+class OptionPreferences { public: bool useCameraInReplays(void); };
+
+// ?useCameraInReplays@OptionPreferences@@QAE_NXZ
+__declspec(naked) bool OptionPreferences::useCameraInReplays(void)
+{
+__asm {
+		_emit 051h
+		_emit 056h
+		_emit 057h
+		_emit 08Bh
+		_emit 0F1h
+		_emit 068h
+		_emit 0D0h
+		_emit 06Ch
+		_emit 007h
+		_emit 001h
+		_emit 08Dh
+		_emit 04Ch
+		_emit 024h
+		_emit 00Ch
+		_emit 0E8h
+		_emit 0ADh
+		_emit 082h
+		_emit 07Fh
+		_emit 000h
+		_emit 08Dh
+		_emit 044h
+		_emit 024h
+		_emit 008h
+		_emit 083h
+		_emit 0C6h
+		_emit 004h
+		_emit 050h
+		_emit 08Bh
+		_emit 0CEh
+		_emit 0E8h
+		_emit 08Ah
+		_emit 0A5h
+		_emit 0F7h
+		_emit 0FFh
+		_emit 08Dh
+		_emit 04Ch
+		_emit 024h
+		_emit 008h
+		_emit 08Bh
+		_emit 0F8h
+		_emit 0E8h
+		_emit 013h
+		_emit 070h
+		_emit 07Fh
+		_emit 000h
+		_emit 03Bh
+		_emit 03Eh
+		_emit 075h
+		_emit 007h
+		_emit 083h
+		_emit 0C8h
+		_emit 0FFh
+		_emit 05Fh
+		_emit 05Eh
+		_emit 059h
+		_emit 0C3h
+		_emit 051h
+		_emit 089h
+		_emit 064h
+		_emit 024h
+		_emit 00Ch
+		_emit 08Bh
+		_emit 0CCh
+		_emit 083h
+		_emit 0C7h
+		_emit 014h
+		_emit 057h
+		_emit 0E8h
+		_emit 018h
+		_emit 072h
+		_emit 07Fh
+		_emit 000h
+		_emit 08Bh
+		_emit 00Dh
+		_emit 0ACh
+		_emit 0D5h
+		_emit 02Eh
+		_emit 001h
+		_emit 0E8h
+		_emit 04Dh
+		_emit 03Ch
+		_emit 0F9h
+		_emit 0FFh
+		_emit 05Fh
+		_emit 05Eh
+		_emit 059h
+		_emit 0C3h
+	}
+}
+
